refactor: reduce duplicated service code
